+++ src/components/Link/LinkView.tsx
@@ -3,61 +3,6 @@ import styled from '@emotion/styled';
 
 import Logo from '@/assets/Logo.svg';
 
-const Container = styled.div`
-  display: flex;
-  width: 992px;
-  background-color: #fafbff;
-  height: 600px;
-  justify-content: center;
-  border-radius: 10px;
-  border-width: 1px;
-  border-color: #e7e7e7;
-`;
-
-const StyledImg = styled.img`
-  width: 192px;
-  height: 64px;
-  margin-bottom: 48px;
-  margin-top: 40px;
-`;
-
-const LinkInfo = styled.div`
-  width: 400px;
-  height: 376px;
-  margin-right: 40px;
-`;
-
-const Divider = styled.div`
-  width: 1px;
-  height: 536px;
-  background-color: #e7e7e7;
-  margin: 20px 0;
-`;
-
-const MemoBox = styled.div`
-  width: 400px;
-  margin-left: 40px;
-`;
-
-const CharacterCount = styled.p`
-  position: fixed;
-  bottom: 130px;
-  right: 110px;
-  font-size: 12px;
-  color: #777;
-`;
-
-const Title = styled.div`
-  font-size: 14px;
-  float: left;
-  margin-top: 18px;
-  height: 40px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  width: 300px;
-`;
-
 interface LinkViewProps {
   category: string;
   linkUrl: string;
@@ -139,3 +84,58 @@ function LinkViewComponent({
 }
 
 export default LinkViewComponent;
+
+const Container = styled.div`
+  display: flex;
+  width: 992px;
+  background-color: #fafbff;
+  height: 600px;
+  justify-content: center;
+  border-radius: 10px;
+  border-width: 1px;
+  border-color: #e7e7e7;
+`;
+
+const StyledImg = styled.img`
+  width: 192px;
+  height: 64px;
+  margin-bottom: 48px;
+  margin-top: 40px;
+`;
+
+const LinkInfo = styled.div`
+  width: 400px;
+  height: 376px;
+  margin-right: 40px;
+`;
+
+const Divider = styled.div`
+  width: 1px;
+  height: 536px;
+  background-color: #e7e7e7;
+  margin: 20px 0;
+`;
+
+const MemoBox = styled.div`
+  width: 400px;
+  margin-left: 40px;
+`;
+
+const CharacterCount = styled.p`
+  position: fixed;
+  bottom: 130px;
+  right: 110px;
+  font-size: 12px;
+  color: #777;
+`;
+
+const Title = styled.div`
+  font-size: 14px;
+  float: left;
+  margin-top: 18px;
+  height: 40px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  width: 300px;
+`;
